Time: 5 ms (63.92%), Space: 46.5 MB (32.36%) - LeetHub

+++ 1513-number-of-substrings-with-only-1s/1513-number-of-substrings-with-only-1s.java
@@ -16,7 +16,7 @@ class Solution {
         }
 
         if(i != j){
-            int len = j - i;
+            long len = j - i;
             ans = (ans + len * (len + 1) / 2);
         }
